[REFACTOR] SafeStoreService, MailService 에 메소드 분배
